+++ _unittests/ut_torch_models/test_llm_phi2.py
@@ -1,14 +1,8 @@
 import unittest
 import torch
-from onnx_diagnostic.ext_test_case import (
-    ExtTestCase,
-    ignore_warnings,
-    requires_transformers,
-    requires_python,
-)
+from onnx_diagnostic.ext_test_case import ExtTestCase, ignore_warnings, requires_transformers
 from onnx_diagnostic.torch_models.llms import get_phi2
 from onnx_diagnostic.helpers import string_type
-from onnx_diagnostic.torch_export_patches import bypass_export_some_errors
 
 
 class TestLlmPhi(ExtTestCase):
@@ -28,23 +22,6 @@ class TestLlmPhi(ExtTestCase):
         )
         ep = torch.export.export(model, (), kwargs=inputs, dynamic_shapes=ds)
         assert ep
-
-    @ignore_warnings(UserWarning)
-    @requires_python((3, 12))
-    def test_export_phi2_2_bypassed(self):
-        data = get_phi2(num_hidden_layers=2)
-        model, inputs, ds = data["model"], data["inputs"], data["dynamic_shapes"]
-        self.assertEqual(
-            {"attention_mask", "past_key_values", "input_ids", "position_ids"}, set(inputs)
-        )
-        with bypass_export_some_errors(patch_transformers=True) as modificator:
-            inputs = modificator(inputs)
-            ep = torch.export.export(model, (), kwargs=inputs, dynamic_shapes=ds, strict=False)
-            assert ep
-        with bypass_export_some_errors(patch_transformers=True) as modificator:
-            inputs = modificator(inputs)
-            ep = torch.export.export(model, (), kwargs=inputs, dynamic_shapes=ds, strict=False)
-            assert ep
 
 
 if __name__ == "__main__":
